requirements.txt
# Swictation - Voice Dictation System for Sway/Wayland
# Successfully tested on Python 3.13.3 with NVIDIA RTX A1000 (CUDA 12.9)
# Installation: pip3 install --break-system-packages -r requirements.txt

# Core ML Framework (pre-installed by user)
torch==2.8.0+cu129
torchaudio>=2.1.0

# NVIDIA NeMo for Canary-1B-Flash STT
nemo_toolkit[asr]==2.5.2

# Audio Processing
sounddevice==0.5.3
soundfile==0.13.1
librosa==0.11.0

# NeMo Dependencies (resolved for Python 3.13)
# Transformers ecosystem
huggingface-hub==0.36.0  # Must be <1.0 for transformers compatibility
tokenizers==0.22.1
accelerate==1.11.0
safetensors==0.6.2
transformers==4.57.1

# Neural network training
pytorch-lightning==2.5.5
lightning-utilities==0.15.2
torchmetrics>=3.0.0

# Data processing
datasets==4.3.0
pandas==2.3.3
pyarrow==22.0.0
dill==0.4.0
multiprocess==0.70.16
xxhash==3.6.0

# Audio & speech processing
lhotse==1.31.1
intervaltree==3.1.0
cytoolz==1.1.0
lilcom==1.8.1
tabulate==0.9.0

# Text error analysis
texterrors==1.0.9
Levenshtein==0.27.1
termcolor==3.2.0
plac==1.4.5

# Configuration management
fsspec==2024.12.0  # Pinned for NeMo compatibility
omegaconf>=2.3.0
hydra-core>=1.3.0
pyyaml>=6.0
toml==0.10.2

# ONNX export
onnx>=1.15.0
ml_dtypes==0.5.3

# Logging & telemetry
tensorboard==2.20.0
tensorboard-data-server==0.7.2
werkzeug==3.1.3
grpcio==1.76.0
absl-py==2.3.1

# NV One Logger
StrEnum==0.4.15
overrides==7.7.0
setuptools==80.9.0

# Graph utilities
graphviz==0.21
libcst==1.8.5

# Annotation tools
pyannote.core==6.0.1
pyannote.metrics==4.0.0
pyannote-database==6.1.0
sortedcontainers==2.4.0

# Build dependencies
pybind11==3.0.1
rapidfuzz>=3.9.0

# Text Processing Pipeline (TODO: for future midstream integration)
# midstream>=0.1.0

# IPC and system integration (TODO: for daemon implementation)
# dbus-python>=1.3.2
